Move base map service logic to new crud map repository
